+++ contrib/electrum-plugin/__init__.py
@@ -1,0 +1,5 @@
+from electrum.i18n import _
+
+fullname = _('bwt')
+description = _('Connect with Bitcoin Core via bwt')
+available_for = ['qt', 'cmdline']