+++ contrib/electrum-plugin/cmdline.py
@@ -1,0 +1,7 @@
+from .bwt import BwtPlugin
+
+class Plugin(BwtPlugin):
+  pass
+
+# TODO: find a way to get the list of loaded wallets when starting up,
+# similarly to what the qt mode does using the init_qt hook.